TODO: use Event4J instead of platform event implementation
