+++ core/src/main/java/net/okocraft/okochat/core/event/LunaChatBungeeBaseEvent.java
@@ -3,39 +3,26 @@
  * @license    LGPLv3
  * @copyright  Copyright ucchy 2020
  */
-package net.okocraft.okochat.core.bukkit.event;
+package net.okocraft.okochat.core.event;
 
 import com.github.ucchyocean.lc3.LunaChat;
 import com.github.ucchyocean.lc3.channel.Channel;
-import org.bukkit.Bukkit;
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
+import net.md_5.bungee.api.plugin.Event;
 
 /**
  * 基底イベントクラス
  * @author ucchy
  */
-public abstract class LunaChatBukkitBaseEvent extends Event {
+public abstract class LunaChatBungeeBaseEvent extends Event {
 
-    private static final HandlerList handlers = new HandlerList();
     protected String channelName;
 
     /**
      * コンストラクタ
      * @param channelName チャンネル名
      */
-    public LunaChatBukkitBaseEvent(String channelName) {
-        super(!Bukkit.isPrimaryThread());
+    public LunaChatBungeeBaseEvent(String channelName) {
         this.channelName = channelName;
-    }
-
-    @Override
-    public HandlerList getHandlers() {
-        return handlers;
-    }
-
-    public static HandlerList getHandlerList() {
-        return handlers;
     }
 
     /**
