+++ core/src/main/java/net/okocraft/okochat/core/event/chat/LunaChatBungeePreChatEvent.java
@@ -3,10 +3,11 @@
  * @license    LGPLv3
  * @copyright  Copyright ucchy 2020
  */
-package net.okocraft.okochat.core.bungee.event;
+package net.okocraft.okochat.core.event.chat;
 
 import com.github.ucchyocean.lc3.channel.Channel;
 import com.github.ucchyocean.lc3.member.ChannelMember;
+import net.okocraft.okochat.core.event.LunaChatBungeeBaseCancellableEvent;
 
 /**
  * チャンネルチャットへの発言前に発生するイベント
